Update models.py to make them compatible with current Ormar and Pydantic
Changes are required to work with the current version of ormar (i.e. migrate to ormar 0.20 and above). From the ormar documentation:

"Instead of defining a Meta class now each of the ormar models require an ormar_config parameter that is an instance of the OrmarConfig class. Note that the attribute must be named ormar_config and be an instance of the config class."
https://collerek.github.io/ormar/latest/migration/

All the "Meta" nested classes have been removed, and instead the fields implemented in OrmarConfig objects.

Additionally, a Config class, which was previous nested in the User class, has also been removed. The existence of this resulted in a Pydantic error:
"pydantic.errors.PydanticUserError: "Config" and "model_config" cannot be used together".

This class only held one value:
class Config:
   use_enum_values = True

As far as I know, the above Config class is not used anywhere. When I searched the project for any references to "use_enum_values", I did not find anything. However, it is possible that I am misunderstanding its use.

Thank you for reading this. These changes are required to build the project with current libraries. Hopefully they will help new developers, and old developers who do a clean install.